--- test/e2e/playwright/swap/helpers/utils.ts
@@ -1,1 +1,0 @@
-export const wait = (n) => new Promise((resolve) => setTimeout(resolve, n));
